VMM/CPUM-armv8: Allow reading/writing the debug control/value registers through CPUMQueryGuestSysReg/CPUMSetGuestSysReg, bugreef:10393 [windows build fix]
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@107032 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/VMM/VMMR3/CPUM-armv8.cpp
@@ -94,7 +94,7 @@
     RINT(a_uMsr, a_uMsr, kCpumSysRegRdFn_FixedValue, kCpumSysRegWrFn_IgnoreWrite, 0, a_uValue, UINT64_MAX, 0, a_szName)
 /** Read/Write value from/to CPUMCTX. */
 #define MVRW(a_uMsr, a_szName, a_offCpum) \
-    RINT(a_uMsr, a_uMsr, kCpumSysRegRdFn_ReadCpumOff, kCpumSysRegWrFn_WriteCpumOff, a_offCpum, 0, UINT64_MAX, 0, a_szName)
+    RINT(a_uMsr, a_uMsr, kCpumSysRegRdFn_ReadCpumOff, kCpumSysRegWrFn_WriteCpumOff, (uint32_t)a_offCpum, 0, UINT64_MAX, 0, a_szName)
 
 
 /*********************************************************************************************************************************
